+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8] - 2026-02-09
+
+### Added
+
+- Added name highlighting to logger output
+- Added support for extra keys in logger output
+- Added get_parameter method to ETLStage class
+
 ## [0.1.7] - 2026-02-08
 
 ### Added
@@ -129,6 +137,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python 3.9+ support
 - Apache-2.0 license
 
+[0.1.8]: https://github.com/alkndoom/stagecraft/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/alkndoom/stagecraft/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/alkndoom/stagecraft/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/alkndoom/stagecraft/compare/v0.1.4...v0.1.5